perf(update): isolate managed environment scoping from preflight
