+++ src/cli/update-cli/update-command-service-env.ts
@@ -63,6 +63,38 @@ export function resolveServiceRefreshEnv(
   return resolvedEnv;
 }
 
+/** Run one update phase under the managed Gateway's authoritative environment. */
+export async function withOwnedManagedUpdateEnv<T>(
+  env: NodeJS.ProcessEnv | undefined,
+  run: () => Promise<T>,
+): Promise<T> {
+  if (!env) {
+    return await run();
+  }
+  // Update finalization is a single serialized CLI phase. Some plugin/config owners still read
+  // process.env, so switch the complete phase atomically and restore the caller afterward.
+  const previousEnv = { ...process.env };
+  for (const key of Object.keys(process.env)) {
+    delete process.env[key];
+  }
+  // A caller may pass process.env itself; clearing it must not erase the supplied scope.
+  const phaseEnv = env === process.env ? previousEnv : env;
+  for (const [key, value] of Object.entries(phaseEnv)) {
+    // Node stringifies undefined on assignment; unset selectors must remain absent.
+    if (value !== undefined) {
+      process.env[key] = value;
+    }
+  }
+  try {
+    return await run();
+  } finally {
+    for (const key of Object.keys(process.env)) {
+      delete process.env[key];
+    }
+    Object.assign(process.env, previousEnv);
+  }
+}
+
 export async function withUpdateInProgressEnv<T>(
   invocationCwd: string | undefined,
   run: () => Promise<T>,
